Extract dynmaic postgres url function to one utility function
