+++ api/core/config.py
@@ -29,45 +29,6 @@ class Settings(BaseSettings):
         raise ValueError(v)
 
     PROJECT_NAME: str = "Twitter Clone Project"
-    # SENTRY_DSN: Optional[HttpUrl] = None
-
-    # @validator("SENTRY_DSN", pre=True)
-    # def sentry_dsn_can_be_blank(cls, v: str) -> Optional[str]:
-    #     if len(v) == 0:
-    #         return None
-    #     return v
-
-    # POSTGRES_SERVER: str
-    # POSTGRES_USER: str
-    # POSTGRES_PASSWORD: str
-    # POSTGRES_DB: str
-    # SQLALCHEMY_DATABASE_URI: Optional[PostgresDsn] = None
-
-    # @validator("SQLALCHEMY_DATABASE_URI", pre=True)
-    # def assemble_db_connection(cls, v: Optional[str], values: Dict[str, Any]) -> Any:
-    #     if isinstance(v, str):
-    #         return v
-    #     return PostgresDsn.build(
-    #         scheme="postgresql",
-    #         user=values.get("POSTGRES_USER", ""),
-    #         password=values.get("POSTGRES_PASSWORD", ""),
-    #         host=values.get("POSTGRES_SERVER", ""),
-    #         path=f"/{values.get('POSTGRES_DB', '') or ''}",
-    #     )
-
-    # SMTP_TLS: bool = True
-    # SMTP_PORT: Optional[int] = None
-    # SMTP_HOST: Optional[str] = None
-    # SMTP_USER: Optional[str] = None
-    # SMTP_PASSWORD: Optional[str] = None
-    # EMAILS_FROM_EMAIL: Optional[EmailStr] = None
-    # EMAILS_FROM_NAME: Optional[str] = None
-
-    # @validator("EMAILS_FROM_NAME")
-    # def get_project_name(cls, v: Optional[str], values: Dict[str, Any]) -> str:
-    #     if not v:
-    #         return values.get("PROJECT_NAME", "")
-    #     return v
 
     EMAIL_RESET_TOKEN_EXPIRE_HOURS: int = 48
     EMAIL_TEMPLATES_DIR: str = "/app/app/email-templates/build"
@@ -92,3 +53,21 @@ class Settings(BaseSettings):
 
 
 settings = Settings()
+
+
+def get_db_connection_url():
+    env = os.environ.get("ENV")
+    if not env:
+        return ""
+
+    if env == "localhost-development":
+        return os.environ.get("LOCAL_POSTGRES_URL")
+
+    elif env == "development":
+        return os.environ.get("LOCAL_DOCKER_INTERNAL_POSTGRES_URL")
+
+    elif env == "staging":
+        return os.environ.get("STAGING_POSTGRES_URL")
+
+    elif env == "production":
+        return os.environ.get("PRODUCTION_POSTGRES_URL")
